backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.0.0
python-decouple==3.8
django-filter==23.3
opencv-python>=4.8.0

# SQL Server database packages
mssql-django==1.4
pyodbc==4.0.39